+++ packages/frontend-app-api/src/tree/instantiateAppNodeTree.test.ts
@@ -27,6 +27,7 @@ import {
   createExtensionBlueprint,
   createExtensionDataRef,
   createExtensionInput,
+  createFrontendPlugin,
 } from '@backstage/frontend-plugin-api';
 import {
   createAppNodeInstance,
@@ -59,7 +60,7 @@ function makeSpec<TConfig, TConfigInput>(
     attachTo: extension.attachTo,
     disabled: extension.disabled,
     extension: extension as Extension<unknown, unknown>,
-    plugin: undefined,
+    plugin: createFrontendPlugin({ pluginId: 'app' }),
     ...spec,
   };
 }