+++ packages/frontend-app-api/src/tree/resolveAppNodeSpecs.ts
@@ -14,7 +14,11 @@
  * limitations under the License.
  */
 
-import { Extension, FrontendFeature } from '@backstage/frontend-plugin-api';
+import {
+  createFrontendPlugin,
+  Extension,
+  FrontendFeature,
+} from '@backstage/frontend-plugin-api';
 import { ExtensionParameters } from './readAppExtensionsConfig';
 import { AppNodeSpec } from '@backstage/frontend-plugin-api';
 import { OpaqueFrontendPlugin } from '@internal/frontend';
@@ -87,6 +91,12 @@ export function resolveAppNodeSpecs(options: {
     );
   }
 
+  const appPlugin =
+    plugins.find(plugin => plugin.id === 'app') ??
+    createFrontendPlugin({
+      pluginId: 'app',
+    });
+
   const configuredExtensions = [
     ...pluginExtensions.map(({ plugin, ...extension }) => {
       const internalExtension = toInternalExtension(extension);
@@ -106,8 +116,8 @@ export function resolveAppNodeSpecs(options: {
       return {
         extension: internalExtension,
         params: {
-          source: undefined,
-          plugin: undefined,
+          source: appPlugin,
+          plugin: appPlugin,
           attachTo: internalExtension.attachTo,
           disabled: internalExtension.disabled,
           config: undefined as unknown,