+++ packages/frontend-defaults/src/createApp.tsx
@@ -22,6 +22,7 @@ import {
   FrontendFeature,
   FrontendFeatureLoader,
 } from '@backstage/frontend-plugin-api';
+import { Progress } from '@backstage/core-components';
 // eslint-disable-next-line @backstage/no-relative-monorepo-imports
 import { defaultConfigLoaderSync } from '../../core-app-api/src/app/defaultConfigLoader';
 // eslint-disable-next-line @backstage/no-relative-monorepo-imports
@@ -42,21 +43,63 @@ import { resolveAsyncFeatures } from './resolution';
  * @public
  */
 export interface CreateAppOptions {
-  features?: (FrontendFeature | FrontendFeatureLoader)[];
-  configLoader?: () => Promise<{ config: ConfigApi }>;
-  bindRoutes?(context: { bind: CreateAppRouteBinder }): void;
   /**
-   * The component to render while loading the app (waiting for config, features, etc)
-   *
-   * Is the text "Loading..." by default.
-   * If set to "null" then no loading fallback component is rendered.   *
+   * The list of features to load.
    */
-  loadingComponent?: ReactNode;
-  extensionFactoryMiddleware?:
-    | ExtensionFactoryMiddleware
-    | ExtensionFactoryMiddleware[];
-  pluginInfoResolver?: FrontendPluginInfoResolver;
-  flags?: { allowUnknownExtensionConfig?: boolean };
+  features?: (FrontendFeature | FrontendFeatureLoader)[];
+
+  /**
+   * Allows for the binding of plugins' external route refs within the app.
+   */
+  bindRoutes?(context: { bind: CreateAppRouteBinder }): void;
+
+  /**
+   * Advanced, more rarely used options.
+   */
+  advanced?: {
+    /**
+     * If set to true, the system will silently accept and move on if
+     * encountering config for extensions that do not exist. The default is to
+     * reject such config to help catch simple mistakes.
+     *
+     * This flag can be useful in some scenarios where you have a dynamic set of
+     * extensions enabled at different times, but also increases the risk of
+     * accidentally missing e.g. simple typos in your config.
+     */
+    allowUnknownExtensionConfig?: boolean;
+
+    /**
+     * Sets a custom config loader, replacing the builtin one.
+     *
+     * This can be used e.g. if you have the need to source config out of custom
+     * storages.
+     */
+    configLoader?: () => Promise<{ config: ConfigApi }>;
+
+    /**
+     * Applies one or more middleware on every extension, as they are added to
+     * the application.
+     *
+     * This is an advanced use case for modifying extension data on the fly as
+     * it gets emitted by extensions being instantiated.
+     */
+    extensionFactoryMiddleware?:
+      | ExtensionFactoryMiddleware
+      | ExtensionFactoryMiddleware[];
+
+    /**
+     * The element to render while loading the app (waiting for config, features, etc).
+     *
+     * This is the `<Progress />` component from `@backstage/core-components` by default.
+     * If set to `null` then no loading fallback element is rendered at all.
+     */
+    loadingElement?: ReactNode;
+
+    /**
+     * Allows for customizing how plugin info is retrieved.
+     */
+    pluginInfoResolver?: FrontendPluginInfoResolver;
+  };
 }
 
 /**
@@ -67,14 +110,14 @@ export interface CreateAppOptions {
 export function createApp(options?: CreateAppOptions): {
   createRoot(): JSX.Element;
 } {
-  let suspenseFallback = options?.loadingComponent;
+  let suspenseFallback = options?.advanced?.loadingElement;
   if (suspenseFallback === undefined) {
-    suspenseFallback = 'Loading...';
+    suspenseFallback = <Progress />;
   }
 
   async function appLoader() {
     const config =
-      (await options?.configLoader?.().then(c => c.config)) ??
+      (await options?.advanced?.configLoader?.().then(c => c.config)) ??
       ConfigReader.fromConfigs(
         overrideBaseUrlConfigs(defaultConfigLoaderSync()),
       );
@@ -87,12 +130,10 @@ export function createApp(options?: CreateAppOptions): {
     });
 
     const app = createSpecializedApp({
-      config,
       features: [appPlugin, ...loadedFeatures],
+      config,
       bindRoutes: options?.bindRoutes,
-      extensionFactoryMiddleware: options?.extensionFactoryMiddleware,
-      pluginInfoResolver: options?.pluginInfoResolver,
-      flags: options?.flags,
+      advanced: options?.advanced,
     });
 
     const rootEl = app.tree.root.instance!.getData(
@@ -102,9 +143,10 @@ export function createApp(options?: CreateAppOptions): {
     return { default: () => rootEl };
   }
 
+  const LazyApp = lazy(appLoader);
+
   return {
     createRoot() {
-      const LazyApp = lazy(appLoader);
       return (
         <Suspense fallback={suspenseFallback}>
           <LazyApp />