+++ packages/ui/src/components/HeaderPage/HeaderPage.tsx
@@ -18,9 +18,12 @@ import type { HeaderPageProps } from './types';
 import { Menu } from '../Menu';
 import { Text } from '../Text';
 import { ButtonIcon } from '../ButtonIcon';
-import { RiMore2Line } from '@remixicon/react';
+import { RiArrowRightSLine, RiMore2Line } from '@remixicon/react';
 import { Tabs, TabList, Tab } from '../Tabs';
 import { useStyles } from '../../hooks/useStyles';
+import { Container } from '../Container';
+import { Link } from '../Link';
+import { Fragment } from 'react/jsx-runtime';
 
 /**
  * A component that renders a header page.
@@ -28,15 +31,33 @@ import { useStyles } from '../../hooks/useStyles';
  * @public
  */
 export const HeaderPage = (props: HeaderPageProps) => {
-  const { title, menuItems, tabs, customActions } = props;
+  const { title, menuItems, tabs, customActions, breadcrumbs } = props;
   const { classNames } = useStyles('HeaderPage');
 
   return (
-    <div className={classNames.root}>
+    <Container className={classNames.root}>
       <div className={classNames.content}>
-        <Text variant="title-small" weight="bold" as="h2">
-          {title}
-        </Text>
+        <div className={classNames.breadcrumbs}>
+          {breadcrumbs &&
+            breadcrumbs.map(breadcrumb => (
+              <Fragment key={breadcrumb.label}>
+                <Link
+                  href={breadcrumb.href}
+                  variant="title-small"
+                  weight="bold"
+                  color="secondary"
+                  truncate
+                  style={{ maxWidth: '240px' }}
+                >
+                  {breadcrumb.label}
+                </Link>
+                <RiArrowRightSLine size={16} color="var(--bui-fg-secondary)" />
+              </Fragment>
+            ))}
+          <Text variant="title-small" weight="bold" as="h2">
+            {title}
+          </Text>
+        </div>
         <div className={classNames.controls}>
           {customActions}
           {menuItems && (
@@ -87,6 +108,6 @@ export const HeaderPage = (props: HeaderPageProps) => {
           </Tabs>
         </div>
       )}
-    </div>
+    </Container>
   );
 };